refactor: Change package-name for new model Review and Comment
